+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# TBD - 2024-07-30
+* Added synchronous event processor 
+
 # 2.2.1 - 2024-07-29
 * Updated key decryption & library requirements